requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bokeh=3.2.1=py311h92b7b1e_0
bottleneck=1.3.5=py311hbed6279_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py311h5eee18b_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.7.22=hbcca054_0
catalogue=2.0.10=py311h38be061_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py311h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=unix_pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_0
colorcet=3.0.1=py311h06a4308_0
comm=0.1.2=py311h06a4308_0
commonmark=0.9.1=py_0
confection=0.1.3=py311h781c19f_1
contourpy=1.0.5=py311hdb19cb5_0
cryptography=41.0.3=py311h130f0dd_0
cycler=0.11.0=pyhd3eb1b0_0
cymem=2.0.6=py311h6a678d5_0
cython-blis=0.7.9=py311hf4808d0_0
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.6.7=py311h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
future=0.18.3=pyhd8ed1ab_0
giflib=5.2.1=h5eee18b_3
holoviews=1.17.1=py311h06a4308_0
idna=3.4=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.25.0=py311h92b7b1e_0
ipython=8.15.0=py311h06a4308_0
jedi=0.18.1=py311h06a4308_1
jinja2=3.1.2=py311h06a4308_0
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h5eee18b_1
jupyter_client=8.1.0=py311h06a4308_0
jupyter_core=5.3.0=py311h06a4308_0
kiwisolver=1.4.4=py311h6a678d5_0
langcodes=3.3.0=pyhd8ed1ab_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
leia-br=0.0.1=pypi_0
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
linkify-it-py=2.0.0=py311h06a4308_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py311h06a4308_0
markdown-it-py=2.2.0=py311h06a4308_1
markupsafe=2.1.1=py311h5eee18b_0
matplotlib-base=3.7.2=py311ha02d727_0
matplotlib-inline=0.1.6=py311h06a4308_0
mdit-py-plugins=0.3.0=py311h06a4308_0
mdurl=0.1.0=py311h06a4308_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
munkres=1.1.4=py_0
murmurhash=1.0.7=py311h6a678d5_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py311h06a4308_0
nltk=3.8.1=pyhd8ed1ab_0
numexpr=2.8.7=py311h65dcdc2_0
numpy=1.26.0=py311h08b1b3b_0
numpy-base=1.26.0=py311hf175353_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
packaging=23.1=py311h06a4308_0
pandas=2.1.1=py311ha02d727_0
panel=1.2.3=py311h06a4308_0
param=1.13.0=py311h06a4308_0
parso=0.8.3=pyhd3eb1b0_0
pathy=0.10.2=pyhd8ed1ab_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.0.1=py311ha6cbd5a_0
pip=23.3=py311h06a4308_0
platformdirs=3.10.0=py311h06a4308_0
preshed=3.0.6=py311h6a678d5_0
prompt-toolkit=3.0.36=py311h06a4308_0
psutil=5.9.0=py311h5eee18b_0
pt-core-news-sm=3.5.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pyct=0.5.0=py311h06a4308_0
pydantic=1.10.12=py311h5eee18b_1
pygments=2.15.1=py311h06a4308_1
pyopenssl=23.2.0=py311h06a4308_0
pyparsing=3.0.9=py311h06a4308_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.0=h7a1cb2a_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.11=2_cp311
pytz=2023.3.post1=py311h06a4308_0
pyviz_comms=2.3.0=py311h06a4308_0
pyyaml=6.0.1=py311h5eee18b_0
pyzmq=25.1.0=py311h6a678d5_0
readline=8.2=h5eee18b_0
regex=2023.10.3=py311h5eee18b_0
requests=2.31.0=py311h06a4308_0
rich=12.6.0=pyhd8ed1ab_0
setuptools=68.0.0=py311h06a4308_0
shellingham=1.5.4=pyhd8ed1ab_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=pyhd8ed1ab_0
spacy=3.5.3=py311h4cb112f_0
spacy-legacy=3.0.12=pyhd8ed1ab_0
spacy-loggers=1.0.5=pyhd8ed1ab_0
sqlite=3.41.2=h5eee18b_0
srsly=2.4.8=py311h6a678d5_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.8.0=hdb19cb5_0
thinc=8.1.10=py311h4cb112f_0
tk=8.6.12=h1ccaba5_0
tornado=6.3.3=py311h5eee18b_0
tqdm=4.65.0=py311h92b7b1e_0
traitlets=5.7.1=py311h06a4308_0
typer=0.7.0=pyhd8ed1ab_0
typing-extensions=4.7.1=py311h06a4308_0
typing_extensions=4.7.1=py311h06a4308_0
tzdata=2023c=h04d1e81_0
uc-micro-py=1.0.1=py311h06a4308_0
urllib3=1.26.16=py311h06a4308_0
vadersentiment=3.3.2=pyhd8ed1ab_0
wasabi=1.1.2=py311h38be061_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py311h06a4308_1
wheel=0.41.2=py311h06a4308_0
wordcloud=1.9.2=py311h5eee18b_0
xyzservices=2022.9.0=py311h06a4308_1
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
